Fix recursion bugs in rewritten physical operators

+++ spark-cypher/src/main/scala/org/opencypher/spark/api/io/CAPSTable.scala
@@ -121,7 +121,8 @@ object SparkCypherTable {
     }
 
     override def limit(items: Long): DataFrameTable = {
-      df.limit(items)
+      // TODO: Unsafe `toInt`
+      df.limit(items.toInt)
     }
 
     override def unionAll(other: DataFrameTable): DataFrameTable = {
